+++ utils/declaredScope.js
@@ -2,9 +2,10 @@
 
 exports.__esModule = true;
 
+/** @type {import('./declaredScope').default} */
 exports.default = function declaredScope(context, name) {
   const references = context.getScope().references;
   const reference = references.find((x) => x.identifier.name === name);
-  if (!reference) { return undefined; }
+  if (!reference || !reference.resolved) { return undefined; }
   return reference.resolved.scope.type;
 };