+++ utils/ignore.js
@@ -7,7 +7,10 @@ const extname = require('path').extname;
 const log = require('debug')('eslint-plugin-import:utils:ignore');
 
 // one-shot memoized
-let cachedSet; let lastSettings;
+/** @type {Set<import('./types').Extension>} */ let cachedSet;
+/** @type {import('./types').ESLintSettings} */ let lastSettings;
+
+/** @type {(context: import('eslint').Rule.RuleContext) => Set<import('./types').Extension>} */
 function validExtensions(context) {
   if (cachedSet && context.settings === lastSettings) {
     return cachedSet;
@@ -18,8 +21,10 @@ function validExtensions(context) {
   return cachedSet;
 }
 
+/** @type {import('./ignore').getFileExtensions} */
 function makeValidExtensionSet(settings) {
   // start with explicit JS-parsed extensions
+  /** @type {Set<import('./types').Extension>} */
   const exts = new Set(settings['import/extensions'] || ['.js']);
 
   // all alternate parser extensions are also valid
@@ -37,6 +42,7 @@ function makeValidExtensionSet(settings) {
 }
 exports.getFileExtensions = makeValidExtensionSet;
 
+/** @type {import('./ignore').default} */
 exports.default = function ignore(path, context) {
   // check extension whitelist first (cheap)
   if (!hasValidExtension(path, context)) { return true; }
@@ -55,7 +61,9 @@ exports.default = function ignore(path, context) {
   return false;
 };
 
+/** @type {import('./ignore').hasValidExtension} */
 function hasValidExtension(path, context) {
-  return validExtensions(context).has(extname(path));
+  // eslint-disable-next-line no-extra-parens
+  return validExtensions(context).has(/** @type {import('./types').Extension} */ (extname(path)));
 }
 exports.hasValidExtension = hasValidExtension;